Moved error/patch index to helpers
