+++ src/helpers/index.ts
@@ -3,4 +3,5 @@ export * from './field'
 export * from './file'
 export * from './folder'
 export * from './help'
+export * from './report'
 export * from './table'
