+++ src/helpers/report.ts
@@ -1,0 +1,28 @@
+import * as types from '../types'
+
+export function createErrorIndex(report?: types.IReport) {
+  const errorIndex: types.IErrorIndex = { header: {}, label: {}, row: {}, cell: {} }
+  if (!report) return errorIndex
+  const errorTask = report.tasks[0]
+  if (!errorTask) return errorIndex
+  for (const error of errorTask.errors) {
+    if (!error.rowNumber && !error.fieldNumber) {
+      const headerKey = '1'
+      errorIndex.header[headerKey] = errorIndex.header[headerKey] || []
+      errorIndex.header[headerKey].push(error)
+    } else if (!error.rowNumber) {
+      const labelKey = `${error.fieldName}`
+      errorIndex.label[labelKey] = errorIndex.label[labelKey] || []
+      errorIndex.label[labelKey].push(error)
+    } else if (!error.fieldNumber) {
+      const rowKey = `${error.rowNumber}`
+      errorIndex.row[rowKey] = errorIndex.row[rowKey] || []
+      errorIndex.row[rowKey].push(error)
+    } else {
+      const cellKey = `${error.rowNumber},${error.fieldName}`
+      errorIndex.cell[cellKey] = errorIndex.cell[cellKey] || []
+      errorIndex.cell[cellKey].push(error)
+    }
+  }
+  return errorIndex
+}
